refactor: move bridge to _bridge for encapsulation
