+++ derive_client/_bridge/__init__.py
@@ -1,0 +1,17 @@
+"""Module for bridging assets to and from Derive."""
+
+from .client import BridgeClient
+from .constants import TARGET_SPEED
+from .enums import ChainID, Currency
+from .models import Address
+from .utils import get_prod_lyra_addresses, get_w3_connection
+
+__all__ = [
+    "BridgeClient",
+    "TARGET_SPEED",
+    "ChainID",
+    "Currency",
+    "Address",
+    "get_prod_lyra_addresses",
+    "get_w3_connection",
+]
