+++ derive_client/_bridge/utils.py
@@ -6,14 +6,14 @@ from web3 import Web3
 from web3.contract import Contract
 from web3.datastructures import AttributeDict
 
-from derive_client.bridge.enums import ChainID, RPCEndPoints
-from derive_client.bridge.models import LyraAddresses
-from derive_client.constants import THIS_REPO_ROOT
+from derive_client._bridge.enums import ChainID, RPCEndPoints
+from derive_client._bridge.models import LyraAddresses
+from derive_client.constants import REPO_ROOT
 
 
 def get_prod_lyra_addresses() -> LyraAddresses:
     """Fetch the socket superbridge JSON data."""
-    prod_lyra_addresses = THIS_REPO_ROOT / "data" / "prod_lyra_addresses.json"
+    prod_lyra_addresses = REPO_ROOT / "data" / "prod_lyra_addresses.json"
     return LyraAddresses(chains=json.loads(prod_lyra_addresses.read_text()))
 
 
@@ -30,7 +30,7 @@ def get_contract(w3: Web3, address: str, abi: list) -> Contract:
 
 
 def get_erc20_contract(w3: Web3, token_address: str) -> Contract:
-    erc20_abi_path = THIS_REPO_ROOT / "data" / "erc20.json"
+    erc20_abi_path = REPO_ROOT / "data" / "erc20.json"
     abi = json.loads(erc20_abi_path.read_text())
     return get_contract(w3=w3, address=token_address, abi=abi)
