+++ derive_client/cli.py
@@ -10,7 +10,7 @@ from dotenv import load_dotenv
 from rich import print
 
 from derive_client.analyser import PortfolioAnalyser
-from derive_client.bridge.enums import ChainID, Currency
+from derive_client._bridge import ChainID, Currency
 from derive_client.derive import DeriveClient
 from derive_client.enums import (
     CollateralAsset,
