+++ derive_client/constants.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 from derive_client.enums import Environment
 
-THIS_REPO_ROOT = Path(__file__).parent.parent
+REPO_ROOT = Path(__file__).parent.parent
 
 PUBLIC_HEADERS = {"accept": "application/json", "content-type": "application/json"}
